feat: add schematic to add a module
