+++ packages/@ama-sdk/generator-sdk/src/generators/shell/templates/base/package.json
@@ -123,5 +123,5 @@
     "npm": ">=6.0.0",
     "yarn": ">=3.1.0"
   },
-  "packageManager": "yarn@3.5.1"
+  "packageManager": "yarn@3.6.0"
 }
