+++ packages/@o3r/core/schematics/add-modules/schema.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "http://json-schema.org/schema",
+  "$id": "NgAddModules",
+  "title": "Add available Otter modules",
+  "description": "ng generate Renovate bot",
+  "properties": {
+  }
+}
