+++ packages/@o3r/core/schematics/add-modules/schema.ts
@@ -1,0 +1,4 @@
+import type { JsonObject } from '@angular-devkit/core';
+
+export interface NgAddModulesSchematicsSchema extends JsonObject {
+}
