+++ packages/@o3r/core/schematics/ng-add/index.ts
@@ -7,13 +7,6 @@ import { lastValueFrom } from 'rxjs';
 import type { PackageJson } from 'type-fest';
 import { NgAddSchematicsSchema } from './schema';
 import { displayModuleList } from '../rule-factories/module-list';
-import type { OTTER_MODULE_KEYWORD as OTTER_MODULE_KEYWORD_TYPE, OTTER_MODULE_SUPPORTED_SCOPES as OTTER_MODULE_SUPPORTED_SCOPES_TYPE } from '@o3r/core';
-
-// TODO: Remove this workaround when #374 is implemented
-const OTTER_MODULE_KEYWORD: typeof OTTER_MODULE_KEYWORD_TYPE = 'otter-module';
-
-// TODO: Remove this workaround when #374 is implemented
-const OTTER_MODULE_SUPPORTED_SCOPES: typeof OTTER_MODULE_SUPPORTED_SCOPES_TYPE = ['otter', 'o3r'];
 
 /**
  * Install dev dependency on your application
@@ -69,7 +62,11 @@ export function ngAdd(options: NgAddSchematicsSchema): Rule {
         const { registerPackageCollectionSchematics } = await import('@o3r/schematics');
         return () => registerPackageCollectionSchematics(corePackageJsonContent)(t, c);
       },
-      (t, c) => displayModuleList(OTTER_MODULE_KEYWORD, OTTER_MODULE_SUPPORTED_SCOPES)(t, c)
+      async (t, c) => {
+        const { OTTER_MODULE_KEYWORD, OTTER_MODULE_SUPPORTED_SCOPES } = await import('@o3r/schematics');
+        const displayModuleListRule = displayModuleList(OTTER_MODULE_KEYWORD, OTTER_MODULE_SUPPORTED_SCOPES);
+        return () => displayModuleListRule(t, c);
+      }
     ])(tree, context);
   };
 }
