--- packages/@o3r/core/src/modules/module.constants.ts
@@ -1,6 +1,0 @@
-/** Keyword defining an Otter compatible module */
-export const OTTER_MODULE_KEYWORD = 'otter-module';
-/** List of whitelisted package scopes for Otter modules */
-export const OTTER_MODULE_SUPPORTED_SCOPES = ['otter', 'o3r'] as const;
-/** Package Keyword to identify the module allowing CMS administration */
-export const OTTER_KEYWORD_CMS = 'otter-cms';
