+++ packages/@o3r/core/src/public_api.ts
@@ -3,4 +3,3 @@ export * from './messages/index';
 export * from './store/index';
 export * from './utils/index';
 export * from './log/index';
-export * from './modules/index';
