+++ packages/@o3r/schematics/src/index.ts
@@ -3,3 +3,4 @@ export * from './rule-factories/index';
 export * from './rules/index';
 export * from './tasks/index';
 export * from './utility/index';
+export * from './modules/index';
