+++ packages/@o3r/schematics/src/modules/index.ts
@@ -1,0 +1,2 @@
+export * from './modules.constants';
+export * from './modules.helpers';
